Struct lists::first::List

pub struct List {
    head: Link,
}

functional programming style:

List a = Empty | Elem a (List a)

This means "A List is either Empty or an Element followed by a List" in functional programming. However, in Rust we don't want to have a layout like functional programming. The previous struct of List will give us a layout like:

[Elem A, ptr] -> (Elem B, ptr) -> (Elem C, ptr) -> (Empty *junk*)<Paste>
[] = Stack
() = Heap

In this case a split off C means:

[Elem A, ptr] -> (Elem B, ptr) -> (Empty *junk*)
[Elem C, ptr] -> (Empty *junk*)

On the other hand, a Rust implementation would expect the layout and split to be like:

layout 2:
[ptr] -> (Elem A, ptr) -> (Elem B, ptr) -> (Elem C, *null*)

split off C:
[ptr] -> (Elem A, ptr) -> (Elem B, *null*)
[ptr] -> (Elem C, *null*)

This results in a implementation:

struct Node {
    elem: i32,
    next: List,
}

pub enum List {
    Empty,
    More(Box<Node>),
}

, which fulfills the following:

Tail of a list never allocates extra junk: check!

enum is in delicious null-pointer-optimized form: check!

All elements are uniformly allocated: check!

However, there is one last problem we need to fix to implement our List. We want to make our List public, but the Node private --- while a public enum will make everything public.

List is a public struct

Fields
head: Link
Methods

impl List

Notes on impl:

Selfis an alias for "that type I wrote at top next to impl". Great for not repeating yourself!

We create an instance of a struct in much the same way we declare it, except instead of providing the types of its fields, we initialize them with values.

We refer to variants of an enum using ::, which is the namespacing operator.

The last expression of a function is implicitly returned. This makes simple functions a little neater. You can still use return to return early like other C-like languages.

Notes on ownership (https://cglab.ca/~abeinges/blah/too-many-lists/book/first-ownership.html):

self - Value

&mut self - mutable reference. The only thing you can't do with an &mut is move the value out with no replacement.

&self - shared reference

pub fn new() -> Self

new() will take no parameter and construct an empty list for us

pub fn push(&mut self, elem: i32)

push() will take a list and an element, and return us a list.

Note that in the implementation, we cannot directly add self.head to be the next element in the new node, This is the only thing we can't do with an &mut --- moveing the value out with no replacement.

pub fn pop(&mut self) -> Option<i32>

pop() will

pub fn pop(&mut self) -> Option<i32> {
    let result;
    match self.head {
        Link::Empty => {
            result = None;

        }
        Link::More(ref node) => {
            result = Some(node.elem);
            self.head = node.next;
        }
    };
    result
}

pub fn pop(&mut self) -> Option<i32> {
    let result;
    match mem::replace(&mut self.head, Link::Empty) {
        Link::Empty => {
            result = None;
        }
        Link::More(node) => {
            result = Some(node.elem);
            self.head = node.next;
        }
    };
    result
}

Box is actually really special in Rust, because it's sufficiently built into the language that the compiler lets you do some stuff that nothing else can do. We actually have been doing one such thing this whole time: DerefMove. Whenever you have a pointer type you can derefence it with * or . to get at its contents. Usually you can get a Deref or maybe even a DerefMut, corresponding to a shared or mutable reference respectively.

However because Box totally owns its contents, you can actually move out of a dereference. This is total magic, because there's no way for any other type to opt into this. There's tons of other cool tricks the compiler knows how to do with Box because it just is Box, but they were all feature-gated at 1.0 pending further design. Ideally Box will be totally user definable in the future.