Struct lists::second::List

pub struct List<T> {
    head: Option<Box<Node<T>>>,
}

Fields
head: Option<Box<Node<T>>>
Methods

impl<T> List<T>

pub fn new() -> Self

Note that there nothing pointy in this method -- we don't need to change anything to make the generic work.

Bask in the Glory that is Self, guardian of refactoring and copy-pasta coding. Also of interest, we don't write List<T> when we construct an instance of list. That part's inferred for us based on the fact that we're returning it from a function that expects a List<T>.

pub fn push(&mut self, elem: T)

Note: Because mem::replace(&mut option, None) is such an incredibly common idiom that Option actually just went ahead and made it a method: take. Thus, before we have

next: mem::replace(&mut self.head, None),

Now, we do

next: self.head.take(),

pub fn pop(&mut self) -> Option<T>

Note: match option { None => None, Some(x) => Some(y) } is such an incredibly common idiom that it was called map. map takes a function to execute on x in the Some(x) to produce the y in Some(y). We could write a proper fn and pass it to map, but we'd much rather write what to do inline.

The way to do this is with a closure. Closures are anonymous functions with an extra super-power: they can refer to local variables outside the closure! This makes them super useful for doing all sorts of conditional logic.

Below is the previous impl.

pub fn pop(&mut self) -> Option<i32> {
    match mem::replace(&mut self.head, None) {
        None => None,
        Some(node) => {
            let node = *node;
            self.head = node.next;
            Some(node.elem)
        }
    }
}

pub fn peek(&self) -> Option<&T>

Note that Rust won't allow the following implementation

pub fn peek(&self) -> Option<&T> {
    self.head.map(|node| {
        &node.elem
    })
}

, becase map takes self by value, which would move the Option out of the thing it's in. Previously this was fine because we had just taken it out, but now we actually want to leave it where it was. The correct way to handle this is with the as_ref method on Option.

Additionally, it demotes the Option to an Option to a reference to its internals. We could do this ourselves with an explicit match but ugh no. It does mean that we need to do an extra derefence to cut through the extra indirection, but thankfully the . operator handles that for us.

pub fn peek_mut(&mut self) -> Option<&mut T>

Mutable version of peek()

Important traits for IterMut<'a, T>
impl<'a, T> Iterator for IterMut<'a, T> type Item = &'a mut T;

pub fn iter_mut(&mut self) -> IterMut<T>

iter_mut()

Important traits for IntoIter<T>
impl<T> Iterator for IntoIter<T> type Item = T;

pub fn into_iter(self) -> IntoIter<T>

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter(&self) -> Iter<T>

Note that this is syntax sugar for

pub fn iter<'a>(&'a self) -> Iter<'a, T> {
    Iter { next: self.head.as_ref().map(|node| &**node)
}

impl<T> List<T>